# swagger/docs/services/identity/proto/identity.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "services/identity/proto/identity.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "Identity"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/identity/sign_in": {
      "get": {
        "operationId": "Identity_SignIn",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/identitySignInResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "password",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "Identity"
        ]
      }
    },
    "/identity/sign_up": {
      "post": {
        "operationId": "Identity_SignUp",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/identitySignUpResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/identitySignUpRequest"
            }
          }
        ],
        "tags": [
          "Identity"
        ]
      }
    }
  },
  "definitions": {
    "identitySignInResponse": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/userUser"
        },
        "jwt": {
          "type": "string"
        }
      }
    },
    "identitySignUpRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "identitySignUpResponse": {
      "type": "object",
      "properties": {
        "isSuccess": {
          "type": "boolean"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "userUser": {
      "type": "object",
      "properties": {
        "uuid": {
          "type": "string",
          "title": "@gotags: bson:\"uuid\""
        },
        "email": {
          "type": "string",
          "title": "@gotags: bson:\"email\""
        },
        "hashPassword": {
          "type": "string",
          "title": "@gotags: bson:\"hash_password\""
        },
        "createdAt": {
          "type": "string",
          "format": "int64",
          "title": "@gotags: bson:\"created_at\""
        },
        "updatedAt": {
          "type": "string",
          "format": "int64",
          "title": "@gotags: bson:\"updated_at\""
        }
      }
    }
  }
}